+++ lib/activemodel_phone.rb
@@ -1,3 +1,3 @@
-
+require 'phony'
 require 'active_model/phone'
 require 'active_model/phone/version'