+++ CHANGELOG.md
@@ -3,5 +3,7 @@
 ## 0.0.1 - Initial Release
 
 * Support for `attr_phone` method.
-* Format number to all numeric, no spaces before validation.
-* Support for `_intl`, `_natl`, and `_cc` accessors.
+* Normalize number to all numeric, no spaces before validation.
+* Support for `_intl`, `_natl`, `_local`, and `_cc` accessors.
+* Support for `:before_normalize` hook.
+* Pass-thru of Phony `:spaces` and `:format` options